split new types off into new files
